Code cleanup, rename routes files
