--- backend/bash/ping_callconvo.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-curl -X POST -H 'content-type:application/json' -d '{"fbId": "1260551614030071", "campaignCallId": "589fd0af121b9bcf8c147e4c", "repIds": ["589f825651f086b33b6b0118"]}' http://localhost:8081/api/start/callConvo
