--- backend/bash/ping_campaigncall.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-curl -X POST -H 'content-type:application/json' http://localhost:8081/api/send/campaignCall/58a93b36e664d31d3dc4f529
