--- backend/bash/ping_campaignupdate.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-curl -X POST -H 'content-type:application/json' http://localhost:8081/api/send/campaignUpdate/58a62539659d106fc3500575
