--- backend/bash/ping_test.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-curl -X POST -H 'content-type:application/json' -d '{"fbId": "1260551614030071"}' http://localhost:8081/api/start/test
